+++ packages/react-vtable/src/table-components/custom/vtable-react-attribute-plugin.ts
@@ -46,7 +46,7 @@ export class VTableReactAttributePlugin extends ReactAttributePlugin {
       container = checkFrozenContainer(graphic);
     }
 
-    if (!(element && ReactDOM && ReactDOM.createRoot)) {
+    if (!(element && ReactDOM && (ReactDOM.createRoot || ReactDOM.render))) {
       return;
     }
     const id = isNil(react.id) ? `${graphic.id ?? graphic._uid}_react` : react.id;
@@ -60,18 +60,36 @@ export class VTableReactAttributePlugin extends ReactAttributePlugin {
       const { wrapContainer, nativeContainer } = this.getWrapContainer(stage, container);
 
       if (wrapContainer) {
-        const root = ReactDOM.createRoot(wrapContainer);
-        root.render(element);
-
         if (!this.htmlMap) {
           this.htmlMap = {};
         }
 
-        this.htmlMap[id] = { root, wrapContainer, nativeContainer, container, renderId: this.renderId, graphic };
+        if (ReactDOM.createRoot) {
+          const root = ReactDOM.createRoot(wrapContainer);
+          root.render(element);
+          this.htmlMap[id] = { root, wrapContainer, nativeContainer, container, renderId: this.renderId, graphic };
+        } else {
+          ReactDOM.render(element, wrapContainer);
+
+          this.htmlMap[id] = {
+            wrapContainer,
+            nativeContainer,
+            container,
+            renderId: this.renderId,
+            unmount: () => {
+              ReactDOM.unmountComponentAtNode(wrapContainer);
+            },
+            graphic
+          };
+        }
       }
     } else {
       // update react element
-      this.htmlMap[id].root.render(element);
+      if (ReactDOM.createRoot) {
+        this.htmlMap[id].root.render(element);
+      } else {
+        ReactDOM.render(element, this.htmlMap[id].wrapContainer);
+      }
     }
 
     if (!this.htmlMap || !this.htmlMap[id]) {
@@ -223,7 +241,7 @@ export class VTableReactAttributePlugin extends ReactAttributePlugin {
 }
 
 function checkFrozenContainer(graphic: IGraphic) {
-  const { col, row, stage } = graphic.parent;
+  const { col, row, stage } = getTargetGroup(graphic);
   let { container } = graphic.attribute.react;
   const { table } = stage as any;
   // deal with react dom container
@@ -253,4 +271,14 @@ function checkFrozenContainer(graphic: IGraphic) {
 
   // graphic.attribute.react.container = container;
   return container;
+}
+
+function getTargetGroup(target: any) {
+  while (target && target.parent) {
+    if (target.name === 'custom-container') {
+      return target;
+    }
+    target = target.parent;
+  }
+  return null;
 }