+++ packages/vtable-export/demo/menu.ts
@@ -49,6 +49,10 @@ export const menus = [
       {
         path: 'list',
         name: 'list-group'
+      },
+      {
+        path: 'list',
+        name: 'list-custom'
       }
     ]
   },