+++ packages/vtable-gantt/src/scenegraph/scenegraph.ts
@@ -191,6 +191,7 @@ export class Scenegraph {
     } as any);
     this.grid.resize();
     this.taskBar.resize();
+    this.dependencyLink.resize();
     this.markLine.refresh();
     this.frameBorder.resize();
   }
@@ -327,6 +328,7 @@ export class Scenegraph {
       let linkedToTaskShowIndex;
       let linkedFromTaskShowIndex;
       let diffY: number;
+      const taskBarStyle = this._gantt.getTaskBarStyle(taskIndex, sub_task_index);
       if (gantt.parsedOptions.tasksShowMode === TasksShowMode.Sub_Tasks_Inline) {
         const new_indexs = getTaskIndexsByTaskY(target.attribute.y + dy, gantt);
         linkedFromTaskShowIndex = linkedFromTaskRecord.index[0];
@@ -345,7 +347,7 @@ export class Scenegraph {
           taskDays: linkedFromTaskTaskDays
         } = gantt.getTaskInfoByTaskListIndex(linkedFromTaskRecord.index[0], linkedFromTaskRecord.index[1]));
 
-        const taskbarHeight = gantt.parsedOptions.taskBarStyle.width;
+        const taskbarHeight = taskBarStyle.width;
         diffY = target.attribute.y + taskbarHeight / 2 - (linkedToTaskShowIndex + 0.5) * gantt.parsedOptions.rowHeight;
       } else if (
         gantt.parsedOptions.tasksShowMode === TasksShowMode.Sub_Tasks_Separate ||
@@ -384,7 +386,7 @@ export class Scenegraph {
           taskDays: linkedFromTaskTaskDays
         } = gantt.getTaskInfoByTaskListIndex(linkedFromTaskRecord.index[0], linkedFromTaskRecord.index[1]));
 
-        const taskbarHeight = gantt.parsedOptions.taskBarStyle.width;
+        const taskbarHeight = taskBarStyle.width;
         diffY = target.attribute.y + taskbarHeight / 2 - (linkedToTaskShowIndex + 0.5) * gantt.parsedOptions.rowHeight;
       } else {
         linkedFromTaskShowIndex = gantt.getTaskShowIndexByRecordIndex(linkedFromTaskRecord.index);
@@ -460,6 +462,7 @@ export class Scenegraph {
       let linkedToTaskShowIndex;
       let linkedFromTaskShowIndex;
       let diffY: number;
+      const taskBarStyle = this._gantt.getTaskBarStyle(taskIndex, sub_task_index);
       if (gantt.parsedOptions.tasksShowMode === TasksShowMode.Sub_Tasks_Inline) {
         const new_indexs = getTaskIndexsByTaskY(target.attribute.y + dy, gantt);
         const beforeRowCountLinkedFrom =
@@ -478,7 +481,7 @@ export class Scenegraph {
           endDate: linkedFromTaskEndDate,
           taskDays: linkedFromTaskTaskDays
         } = gantt.getTaskInfoByTaskListIndex(linkedFromTaskRecord.index[0], linkedFromTaskRecord.index[1]));
-        const taskbarHeight = gantt.parsedOptions.taskBarStyle.width;
+        const taskbarHeight = taskBarStyle.width;
         diffY =
           target.attribute.y + taskbarHeight / 2 - (linkedFromTaskShowIndex + 0.5) * gantt.parsedOptions.rowHeight;
       } else if (
@@ -517,7 +520,7 @@ export class Scenegraph {
           endDate: linkedFromTaskEndDate,
           taskDays: linkedFromTaskTaskDays
         } = gantt.getTaskInfoByTaskListIndex(linkedFromTaskRecord.index[0], linkedFromTaskRecord.index[1]));
-        const taskbarHeight = gantt.parsedOptions.taskBarStyle.width;
+        const taskbarHeight = taskBarStyle.width;
         diffY =
           target.attribute.y + taskbarHeight / 2 - (linkedFromTaskShowIndex + 0.5) * gantt.parsedOptions.rowHeight;
       } else {