+++ packages/vtable/src/event/event.ts
@@ -170,7 +170,11 @@ export class EventManager {
         );
         if (this.table.eventManager.checkCellFillhandle(eventArgsSet)) {
           this.table.fireListeners(TABLE_EVENT_TYPE.DBLCLICK_FILL_HANDLE, {});
-        } else if (this.table._canResizeColumn(resizeCol.col, resizeCol.row) && resizeCol.col >= 0) {
+        } else if (
+          this.table._canResizeColumn(resizeCol.col, resizeCol.row) &&
+          resizeCol.col >= 0 &&
+          !this.table.options.disableDblclickAutoResizeColWidth
+        ) {
           this.table.scenegraph.updateAutoColWidth(resizeCol.col);
           this.table.internalProps._widthResizedColMap.add(resizeCol.col);
           // if (this.table.isPivotChart()) {
@@ -524,6 +528,11 @@ export class EventManager {
       }
       return true;
     }
+    if (this.table.stateManager.isResizeCol()) {
+      // 结束列调整
+      this.table.stateManager.endResizeCol();
+    }
+
     // }
 
     return false;