+++ packages/vtable/src/ts-types/list-table/define/link-define.ts
@@ -14,9 +14,9 @@ export interface ILinkHeaderDefine extends IBasicHeaderDefine {
 
   // 这里也需要定义 和bodyDefine的作用是不一样的
   /** 链接是否可点击跳转 */
-  linkJump?: boolean;
+  linkJump?: boolean | ((args: StylePropertyFunctionArg) => boolean);
   /** 链接是否进行正则检测，如果链接符合url规则才展示成为link。如果配置了模板链接该配置不生效。 */
-  linkDetect?: boolean;
+  linkDetect?: boolean | ((args: StylePropertyFunctionArg) => boolean);
   /** 模板链接地址，如：'https://www.google.com.hk/search?q={name}'，name是数据源属性字段名。 */
   templateLink?: string | FieldGetter;
 
@@ -33,9 +33,9 @@ export interface ILinkColumnBodyDefine extends IBasicColumnBodyDefine {
   // autoWrapText?: boolean;
   // lineClamp?: LineClamp;
   /** 链接是否可点击跳转 */
-  linkJump?: boolean;
+  linkJump?: boolean | ((args: StylePropertyFunctionArg) => boolean);
   /** 链接是否进行正则检测，如果链接符合url规则才展示成为link。如果配置了模板链接该配置不生效。 */
-  linkDetect?: boolean;
+  linkDetect?: boolean | ((args: StylePropertyFunctionArg) => boolean);
   /** 模板链接地址，如：'https://www.google.com.hk/search?q={name}'，name是数据源属性字段名。 */
   templateLink?: string | FieldGetter;